app-api/pyproject.toml
[project]
name = "recruitai-backend"
version = "0.1.0"
description = "Recruit-AI backend following Clean Architecture with FastAPI and PostgreSQL"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.5",
  "uvicorn[standard]>=0.32.0",
  "sqlalchemy>=2.0.36",
  "psycopg2-binary>=2.9.10",
  "pydantic>=2.9.2",
  "pydantic-settings>=2.6.1",
  "email-validator>=2.2.0",
  "boto3>=1.35.67",
  "PyJWT>=2.9.0",
  "python-multipart>=0.0.9",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt>=4.1.2",
  "alembic>=1.13.3"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
  "httpx>=0.27.2",
  "ruff>=0.7.3"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
